Fix fetching of medium-size user avatars.
In the refactoring in ded7baa61b3127014ca587e54fff85074a445d30 that
fixed live-updating of the medium-size avatar data, we started just
fetching the normal-size avatar, not the medium-size avatar.  We fix
this by changing this code path to pass in the user object and
construct the URL using that.

While we're at it, we switch to using the user ID, not the email, to
construct these avatar URLs.