vagrant: Add NFS backend for file synchronization for OSX.
This is intended to fix serious issues with frequent guest crashes on
OSX High Sierra with Virtualbox and the default file synchronization.

Note from Tim: We've tried this idea before, and it worked for Mac,
but caused problems for non-Mac users of our Vagrant setup that
resulted in it being reverted in
93b4e2042fdd308fcf3e585d4409b42dcb6138da.

Hopefully the new conditional will handle things correctly in both Mac
and Linux.